talk-tidbits-backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
aniso8601==9.0.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.2
decorator==5.1.1
exceptiongroup==1.1.1
ffmpeg-python==0.2.0
filelock==3.10.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.39.1
Mako==1.2.4
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.0
numba==0.56.4
numpy==1.23.5
openai==0.27.2
openai-whisper==20230314
packaging==23.0
pluggy==1.0.0
pooch==1.6.0
psycopg2-binary==2.9.5
pycparser==2.21
pytest==7.2.2
pytest-cov==4.0.0
pytest-flask==1.2.0
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.4
SQLAlchemy==2.0.7
sympy==1.11.1
threadpoolctl==3.1.0
tiktoken==0.3.1
tomli==2.0.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
